v2.5
- Обновлена структура
- Исправлена система обработки повторов
