+++ config.txt
@@ -1,1 +1,1 @@
-токен бота
+7155133870:AAH2Xc1DsYR-7W_2Ap1bBWhXwQH4zNDQFNg
