+++ handlers/__init__.py
@@ -1,8 +1,9 @@
 from aiogram import Dispatcher
 
-from handlers import rand_find
+from handlers import rand_find, start
 
 def include_routers(dp: Dispatcher):
     dp.include_routers(
         rand_find.router,
+        start.router
     )
